pyproject.toml
[project]
name = "repository"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate==1.5.2",
    "datasets==3.4.1",
    "torch==2.6.0",
    "transformers==4.49.0",
    "python-dotenv==1.0.0",
]
